harmony: move harmony parsing into a package (#12016)
